refactor: remove export default and rename file
